настройки для сборки npm пакета вынес header в отдельный компонент
