+++ package.json
@@ -1,12 +1,17 @@
 {
   "name": "dochub",
-  "version": "0.1.0",
+  "version": "2.6.2",
   "private": true,
+  "main": "./dist/dochub.js",
+  "files": [
+    "./dist"
+  ],
   "scripts": {
     "lib": "vue-cli-service build --name test --target lib",
     "plugin": "export VUE_APP_DOCHUB_MODE=plugin && vue-cli-service build",
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
+    "build:package": "node vue.lib.js",
     "lint": "vue-cli-service lint",
     "test": "jest --watch",
     "meta-import": "node src/tools/meta/import.js"
